Refactoring components
Added components folders

Added files for components in components folders
