+++ .vs/VSWorkspaceState.json
@@ -1,6 +1,8 @@
 {
   "ExpandedNodes": [
-    ""
+    "",
+    "\\public",
+    "\\src"
   ],
   "SelectedNode": "\\package.json",
   "PreviewInSolutionExplorer": false
